IQSD-1931: Разбитие chat.vue на компоненты
